Facial_Voice_Sense/requirements.txt
# Facial Voice Sense - Python依赖包配置文件
# 使用方法: pip install -r requirements.txt

# ==================== 核心深度学习框架 ====================
tensorflow>=2.10.0
keras>=2.10.0

# ==================== 计算机视觉 ====================
opencv-python>=4.5.0
Pillow>=9.0.0

# ==================== 数值计算 ====================
numpy>=1.21.0
numba>=0.56.0

# ==================== 音频处理 ====================
sounddevice>=0.4.5
soundfile>=0.11.0
librosa>=0.9.0

# ==================== 语音识别 ====================
funasr>=1.0.0

# ==================== 中文处理 ====================
jieba>=0.42.0

# ==================== 配置管理 ====================
hydra-core>=1.2.0
omegaconf>=2.2.0

# ==================== 文件处理 ====================
h5py>=3.7.0
filelock>=3.8.0
fsspec>=2022.0.0

# ==================== 并行处理 ====================
joblib>=1.2.0

# ==================== 音频IO ====================
kaldiio>=0.9.0

# ==================== 网络通信 ====================
grpcio>=1.50.0

# ==================== 解析器生成器 ====================
antlr4-python3-runtime>=4.9.0

# ==================== 模板引擎 ====================
jinja2>=3.1.0

# ==================== JSON查询 ====================
jmespath>=1.0.0

# ==================== 二进制序列化 ====================
msgpack>=1.0.0

# ==================== HTTP客户端 ====================
requests>=2.28.0
urllib3>=1.26.0

# ==================== 终端美化 ====================
colorama>=0.4.5
rich>=12.0.0
pygments>=2.13.0

# ==================== 工具库 ====================
certifi>=2022.0.0
idna>=3.4
packaging>=21.3
pyparsing>=3.0.0
six>=1.16.0
typing-extensions>=4.4.0

# ==================== Google基础库 ====================
absl-py>=1.2.0

# ==================== AST抽象 ====================
gast>=0.5.0

# ==================== 树结构 ====================
optree>=0.2.0

# ==================== Markdown ====================
markdown>=3.4.0

# ==================== 序列化 ====================
protobuf>=3.20.0

# ==================== TensorFlow相关 ====================
tensorboard>=2.10.0
werkzeug>=2.2.0
tensorboard-data-server>=0.6.0
tensorboard-plugin-wit>=1.8.0

# ==================== 其他工具 ====================
decorator>=5.1.0
pooch>=1.7.0
pyyaml>=6.0
charset-normalizer>=2.1.0